+++ lib/stdio/generator.js
@@ -1,6 +1,6 @@
 import {generatorsToTransform} from './transform.js';
 import {getEncodingTransformGenerator} from './encoding-transform.js';
-import {getLinesGenerator} from './lines.js';
+import {getLinesGenerator} from './split.js';
 import {pipeStreams} from './pipeline.js';
 import {isGeneratorOptions, isAsyncGenerator} from './type.js';
 import {getValidateTransformReturn} from './validate.js';
