+++ lib/stdio/split.js
