+++ test/stdio/split.js
